Revert ":recycle: Use our own canvas component"
