+++ frontend/.eslintrc
@@ -57,8 +57,8 @@
     "@typescript-eslint/interface-name-prefix": 0,
     "complexity": "off",
     "max-lines": ["error", 300],
-    "max-depth": ["error", 4],
-    "max-params": ["error", 10],
+    "max-depth": ["error", 3],
+    "max-params": ["error", 4],
     "risxss/catch-potential-xss-react": "error"
   }
 }
