+++ frontend/package.json
@@ -18,6 +18,7 @@
     "lodash": "^4.17.15",
     "lz-string": "^1.4.4",
     "react": "^16.12.0",
+    "react-canvas-draw": "https://github.com/embiem/react-canvas-draw#develop",
     "react-dom": "^16.12.0",
     "react-intl": "^4.4",
     "react-redux": "^7.1.3",
@@ -93,6 +94,7 @@
     "@types/lodash": "^4.14.149",
     "@types/node": "^10.12.18",
     "@types/react": "^16.9.15",
+    "@types/react-canvas-draw": "^1.1.0",
     "@types/react-dom": "^16.9.4",
     "@types/react-intl": "^3.0.0",
     "@types/react-redux": "^7.1.5",
