--- frontend/src/components/Canvas/CanvasDraw/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CanvasDraw';
