--- frontend/src/components/Canvas/CanvasRecap/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CanvasRecap';
