+++ frontend/src/components/CanvasActions/CanvasActions.style.ts
@@ -1,10 +1,15 @@
-import deleteIcon from 'assets/delete.svg';
-import undoIcon from 'assets/undo.svg';
 import styled from 'styled-components';
+
+import undoIcon from 'assets/undo.svg';
+import deleteIcon from 'assets/delete.svg';
 
 export const CanvasActionsContainer = styled.div`
   display: flex;
   flex-direction: column;
+  position: absolute;
+  top: 16px;
+  right: 16px;
+  z-index: 25;
 `;
 
 CanvasActionsContainer.displayName = 'CanvasActionsContainer';
