+++ frontend/src/components/CanvasActions/CanvasActions.tsx
