+++ frontend/src/components/CanvasActions/index.ts
